Controles de consistencia en los movimientos de caja. Control de consistencia en el borrado de cheques. Control de consistencia en el borrado de bancos. Mejora en la visualizacion de saldos en las compras.

+++ trivoliSwimming/config/cheques_con_04.asp
@@ -19,15 +19,15 @@ l_id = request.Form("cabnro")
 Set l_rs = Server.CreateObject("ADODB.RecordSet")
 Set l_cm = Server.CreateObject("ADODB.Command")
 
-'l_sql = "SELECT id"
-'l_sql = l_sql & " FROM calendarios"
-'l_sql  = l_sql  & " WHERE idrecursoreservable = " & l_id
-'rsOpen l_rs, cn, l_sql, 0 
+l_sql = "SELECT id"
+l_sql = l_sql & " FROM cajamovimientos "
+l_sql  = l_sql  & " WHERE idcheque = " & l_id
+rsOpen l_rs, cn, l_sql, 0 
 
-'if not l_rs.eof then
-'	Response.write "Existen calendarios asociados al M&eacute;dico. No se permite eliminar."
-'	l_rs.close
-'else
+if not l_rs.eof then
+	Response.write "Existen movimientos de caja asociados al Cheque. No se permite eliminar."
+	l_rs.close
+else
 	l_sql = "DELETE FROM cheques  WHERE id = " & l_id
 
 	l_cm.activeconnection = Cn
@@ -38,7 +38,7 @@ Set l_cm = Server.CreateObject("ADODB.Command")
 	Set cn = Nothing
 	
 	Response.write "OK"
-'end if
+end if
